refactor(skip): Skip conflicts on bulk upload PE-633
