+++ src/arfs_file_wrapper.ts
@@ -3,6 +3,7 @@ import { extToMime } from 'ardrive-core-js';
 import { basename, join } from 'path';
 import { ByteCount, DataContentType, FileID, FolderID, UnixTime } from './types';
 import { BulkFileBaseCosts, MetaDataBaseCosts } from './ardrive';
+import { EntityNamesAndIds } from './utils/mapper_functions';
 
 type BaseFileName = string;
 type FilePath = string;
@@ -61,6 +62,7 @@ export class ArFSFileToUpload {
 
 	baseCosts?: BulkFileBaseCosts;
 	existingId?: FileID;
+	folderNameConflict = false;
 	hasSameLastModifiedDate = false;
 
 	public gatherFileInfo(): FileInfo {
@@ -107,6 +109,7 @@ export class ArFSFolderToUpload {
 	baseCosts?: MetaDataBaseCosts;
 	existingId?: FolderID;
 	destinationName?: string;
+	fileNameConflict = false;
 
 	constructor(public readonly filePath: FilePath, public readonly fileStats: fs.Stats) {
 		const entitiesInFolder = fs.readdirSync(this.filePath);
@@ -123,6 +126,70 @@ export class ArFSFolderToUpload {
 				// Child is a file, build a new file
 				const childFile = new ArFSFileToUpload(absoluteEntityPath, entityStats);
 				this.files.push(childFile);
+			}
+		}
+	}
+
+	public async checkAndAssignExistingNames(
+		getExistingNamesFn: (parentFolderId: FolderID) => Promise<EntityNamesAndIds>
+	): Promise<void> {
+		if (!this.existingId) {
+			// Folder has no existing ID to check
+			return;
+		}
+
+		const existingEntityNamesAndIds = await getExistingNamesFn(this.existingId);
+
+		for await (const file of this.files) {
+			const baseFileName = file.getBaseFileName();
+
+			const folderNameConflict = existingEntityNamesAndIds.folders.find(
+				({ folderName }) => folderName === baseFileName
+			);
+
+			if (folderNameConflict) {
+				// Folder name cannot conflict with a file name
+				file.folderNameConflict = true;
+				continue;
+			}
+
+			const fileNameConflict = existingEntityNamesAndIds.files.find(({ fileName }) => fileName === baseFileName);
+
+			// Conflicting file name creates a REVISION by default
+			if (fileNameConflict) {
+				file.existingId = fileNameConflict.fileId;
+
+				if (fileNameConflict.lastModifiedDate === file.lastModifiedDate) {
+					// Check last modified date and set to true to resolve upsert conditional
+					file.hasSameLastModifiedDate = true;
+				}
+			}
+		}
+
+		for await (const folder of this.folders) {
+			const baseFolderName = folder.getBaseFileName();
+
+			const fileNameConflict = existingEntityNamesAndIds.files.find(
+				({ fileName }) => fileName === baseFolderName
+			);
+
+			if (fileNameConflict) {
+				// Folder name cannot conflict with a file name
+				this.fileNameConflict = true;
+				continue;
+			}
+
+			const folderNameConflict = existingEntityNamesAndIds.folders.find(
+				({ folderName }) => folderName === baseFolderName
+			);
+
+			// Conflicting folder name uses EXISTING folder by default
+			if (folderNameConflict) {
+				// Assigns existing id for later use
+				folder.existingId = folderNameConflict.folderId;
+
+				// Recurse into existing folder on folder name conflict
+				await folder.checkAndAssignExistingNames(getExistingNamesFn);
 			}
 		}
 	}
